% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ch_api}
\alias{ch_api}
\title{Clubhouse api}
\usage{
ch_api(verb, url, config = list(), ...)
}
\arguments{
\item{verb}{the api verb from the httr package (e.g. GET)}

\item{url}{the full request url}

\item{config}{additional configuration to add to header}

\item{...}{other parameters to pass on to httr api call}
}
\description{
The default base function to allow generic api calls
}
