% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{ch_get_one}
\alias{ch_get_one}
\title{Get endpoint data for one item}
\usage{
ch_get_one(id, endpoint = NULL, url = NULL, config = list(),
  ch_token = get_token())
}
\arguments{
\item{id}{the id for the record you want to retrieve}

\item{endpoint}{data to list. One of "categories",  "epics",
"files", "labels", "linked-files",  "milestones", "projects", "repositories",
"teams",  "iterations"}

\item{url}{Clubhouse API url.  Defaults to API V2}

\item{config}{additional configuration to add to header}

\item{ch_token}{Clubhouse API token.}
}
\description{
Get endpoint data for one item
}
\examples{
\dontrun{
# Retrieve all projects
ch_list_all(endpoint = "projects")

# List epics then retrieve full details about the first one
df <- ch_list_all("epics", response_type = "minimal")
ch_get_one("epics", id = df[ 1, "id"] )
}

}
