% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ch_url}
\alias{ch_url}
\title{Clubhouse url}
\usage{
ch_url(ch_base_url = get_url(), endpoint = NULL, id = NULL,
  ch_token = get_token())
}
\arguments{
\item{ch_base_url}{the base url. Defaults to V2 of API.}

\item{endpoint}{endpoint to retrieve (e.g. project, team, story)}

\item{id}{id of single record to be retrieved using \code{ch_get_one} function}

\item{ch_token}{Clubhouse API token.}
}
\description{
Function to construct urls for Clubhouse api calls
}
\details{
This function constructs the url for each API call. \code{ch_token}
may be input as function parameters or set once with \code{set_token}.
}
