% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterations.R
\name{create_iteration}
\alias{create_iteration}
\title{Create an iteration}
\usage{
create_iteration(
  name,
  start_date,
  end_date,
  description = NA_character_,
  follower_ids = NA_character_,
  base_url = get_url(),
  config = list(),
  ch_token = get_token()
)
}
\arguments{
\item{name}{\code{chr} Required. The name of this Iteration.}

\item{start_date}{\code{chr} Required. The date this Iteration begins, e.g. 2019-07-01.}

\item{end_date}{\code{chr} Required. The date this Iteration ends, e.g. 2019-07-01.}

\item{description}{\code{chr} The description of the Iteration.}

\item{follower_ids}{\code{chr} A vector of UUIDs for any members you want to add as followers}

\item{base_url}{Clubhouse API url.}

\item{config}{additional configuration to add to header}

\item{ch_token}{Clubhouse API token.}
}
\description{
Create an iteration
}
\examples{

\dontrun{
 
create_iteration(
  name = "demo iteration"
  , start_date = "2021-06-01"
  , end_date = "2021-06-30"
  , description = paste0( "# markdown H1 Title \n\n",
                          "This is a description. \n\n ",
                          "THis is a second paragraph."
  )
)
}

}
