% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-settings.R
\name{reset_api}
\alias{reset_api}
\title{Reset API options}
\usage{
reset_api()
}
\description{
Set API url to default and clear API token
}
\examples{

set_token("mytoken") # set token
get_token()  # check value updated
reset_api()  # clear values

}
